op-node/rollup/derive: Reorganize baseBatchStage
Fixes #12567
